pull out events and landing
